+++ src/commands/Fun/meme.js
@@ -5,12 +5,12 @@ module.exports = {
   name: "meme",
   run: async (client, message, args, settings) => {
     if ((await client.isIgnored()) == true) return;
-    message.delete();
+    if (settings.autoDelete == true) message.delete();
     const subReddits = ["LesMemesFrancais", "MemeFrancais", "memes_fr"];
     const random = subReddits[Math.floor(Math.random() * subReddits.length)];
     meme(random, (err, data) => {
       if (err) return console.error(err);
-      message.channel.send({files: [data.url]});
+      message.channel.send({ files: [data.url] });
     });
   },
   cooldown: 5,
