+++ src/commands/Logs/logs.js
@@ -2,7 +2,7 @@ module.exports = {
   name: "logs",
   run: async (client, message, args, settings) => {
     if ((await client.isIgnored()) == true) return;
-    message.delete();
+    if (settings.autoDelete == true) message.delete();
     if (!client.checkPerms("ADMINISTRATOR")) return client.noPerms();
     if (client.isEnabled("logs") == false) return client.moduleDisabled("logs");
     const { MessageEmbed } = require("discord.js");
