+++ src/commands/Moderation/ignorerole.js
@@ -2,7 +2,7 @@ module.exports = {
   name: "ignorerole",
   run: async (client, message, args, settings) => {
     if (await client.isIgnored()) return;
-    message.delete();
+    if (settings.autoDelete == true) message.delete();
     if (!client.checkPerms("ADMINISTRATOR")) return client.noPerms();
     if (client.isEnabled("moderation") == false)
       return client.moduleDisabled("moderation");
@@ -26,7 +26,7 @@ module.exports = {
     }
   },
   cooldown: 5,
-  usage: `prefixname <role_id || role_mention || role_name>`,
+  usage: `prefixname <role>`,
   description:
     "Défini / retire un rôle pour lequel les commandes seront ignorées",
   category: "Moderation",
